requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-environ==0.12.0
django-vite==3.1.0
dnspython==2.8.0
email-validator==2.3.0
Faker==37.12.0
gunicorn==23.0.0
idna==3.11
inertia-django==1.2.0
packaging==25.0
pillow==12.0.0
psycopg==3.2.10
psycopg-binary==3.2.10
pydantic==2.12.0
pydantic_core==2.41.1
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
